+++ forge-annotations/src/main/java/com/kivojenko/spring/forge/annotation/WithJpaRepository.java
@@ -18,4 +18,12 @@ public @interface WithJpaRepository {
      * @return the package name
      */
     String packageName() default "";
+
+    /**
+     * Whether the generated repository should be abstract.
+     * If true, the repository will be abstract, allowing for custom implementation later.
+     *
+     * @return true if the controller should be abstract, false otherwise
+     */
+    boolean makeAbstract() default false;
 }