+++ forge-annotations/src/main/java/com/kivojenko/spring/forge/annotation/WithService.java
@@ -18,4 +18,12 @@ public @interface WithService {
      * @return the package name
      */
     String packageName() default "";
+
+    /**
+     * Whether the generated service should be abstract.
+     * If true, the service will be abstract, allowing for custom implementation later.
+     *
+     * @return true if the controller should be abstract, false otherwise
+     */
+    boolean makeAbstract() default false;
 }