+++ forge-runtime/src/main/java/com/kivojenko/spring/forge/jpa/model/model/JpaEntityModel.java
@@ -229,7 +229,7 @@ public record JpaEntityModel(TypeElement element,
         return element().getEnclosedElements()
                 .stream()
                 .filter(c -> c.getKind() == ElementKind.FIELD)
-                .anyMatch(c -> c.getSimpleName().contentEquals("name"));
+                .anyMatch(c -> c.getSimpleName().contentEquals("name") && isNameType(c.asType()));
     }